refactor: test 코드의 변수값, 헬퍼 메서드를 분리하여 NovelFixture 클래스 생성
